+++ src/GWind/BillingAccount/Control.hs
@@ -1,0 +1,10 @@
+-- ⚠
+module GWind.BillingAccount.Control (reviseBillingAccountsM)
+where
+
+--
+    reviseBillingAccountsM :: IO ()
+    reviseBillingAccountsM  =
+        undefined
+
+--