+++ src/GWind/BillingProject/Domain.hs
@@ -1,0 +1,9 @@
+-- ⚠
+module GWind.BillingProject.Domain (BillingSnake)
+where
+
+    import GWind.Common.Domain (Snake)
+
+    type BillingSnake = Snake
+
+--