+++ src/GWind/Service/Domain.hs
@@ -1,0 +1,7 @@
+-- ⚠
+module GWind.Service.Domain (ServiceNicks)
+where
+
+    type ServiceNicks = String
+
+--