+++ src/GWind/StorageBucket/Control.hs
@@ -1,0 +1,22 @@
+-- ⚠
+module GWind.StorageBucket.Control (reviseStorageBucketsM, createStorageBucketM, deleteStorageBucketM)
+where
+
+    import GWind.Project.Domain (ProjectSnake)
+
+--
+    reviseStorageBucketsM :: ProjectSnake -> IO ()
+    reviseStorageBucketsM _ =
+        undefined
+
+--
+    createStorageBucketM :: ProjectSnake -> IO ()
+    createStorageBucketM _ =
+        undefined
+
+--
+    deleteStorageBucketM :: ProjectSnake -> IO ()
+    deleteStorageBucketM _ =
+        undefined
+
+--